+++ packages/orchestrator/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- `shared` properties and microlcApi extensions are persisted in `compose` applications across mount and unmount cycles
+
 ## [2.0.4] - 2023-03-07
 
 ### Versioning